Extract copy/pasted postgres initialization code
